requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.0
PyYAML>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
pydantic>=2.0.0